size_t

+++ src/CSlim/SlimConnectionHandler.c
@@ -40,7 +40,7 @@ int read_size(SlimConnectionHandler* self)
 	char size[7];
 	int size_i = -1;
 	char colon;
-	memset(size, 0, 7);
+	memset(size, (size_t)0, (size_t)7);
 
 	if (self->recvFunc(self->comLink, size, 6) == 6)
 	{
@@ -55,7 +55,7 @@ int read_size(SlimConnectionHandler* self)
 
 int SlimConnectionHandler_Run(SlimConnectionHandler* self)
 {
-	char * message = (char*)malloc(3);
+	char * message = (char*)malloc((size_t)3);
 	message[0] = 0;
 	int numbytes;
 
@@ -66,12 +66,12 @@ int SlimConnectionHandler_Run(SlimConnectionHandler* self)
 
 	while(1)
 	{
-		int size_i = read_size(self);
+		size_t size_i = read_size(self);
 		if (size_i > 0)
 		{
 			free(message);
-			message = (char*)malloc(size_i + 1);
-			memset(message, 0, size_i + 1);
+			message = (char*)malloc(size_i + (size_t)1);
+			memset(message, 0, size_i + (size_t)1);
 			numbytes = self->recvFunc(self->comLink, message, size_i);
 			if (numbytes != size_i)
 			{
@@ -85,7 +85,7 @@ int SlimConnectionHandler_Run(SlimConnectionHandler* self)
 			//execute and get response
 			char* response_message = self->slimHandlerFunc(self->slimHandler, message);
 			int response_length = strlen(response_message);
-			char * length_buffer = (char *)malloc(8);
+			char * length_buffer = (char *)malloc((size_t)8);
 			sprintf(length_buffer, "%06d:", response_length);
 			int send_result = self->sendFunc(self->comLink, length_buffer, 7);
 			free(length_buffer);
